Create 3D_KDE_With_Shape.R
This file is adds the capability to convert kde's to shapefile.  It is important to note that this now requires three new R libraries (sp, raster, maptools).  There are two options for the final step of writing the shapefile that are both commented out and we can decide which one works best/has better options for the arguments.  I think the raster shapefile() function will work best.  

The conversion works by first converting the kde's points into a spatialDataFrame, adding the countours, and then using raster (or maptools) to convert the spatialDataFrame into a shapefile